Backend/requirements/linux_reqs.txt
asgiref==3.5.2
autopep8==1.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==2.8.0
django-smtp-ssl==1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
importlib-metadata==4.12.0
Markdown==3.3.7
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tzdata==2022.1
Unipath==1.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0